refactor : #86 Entity 기본 생성자 access level 추가
