--- escape-admin/src/main/java/com/sparta/domain/review/dto/ReviewResponseDto.java
@@ -1,4 +1,0 @@
-package com.sparta.domain.review.dto;
-
-public class ReviewResponseDto {
-}
