+++ escape-consumer/src/main/java/com/sparta/domain/consumer/service/ConsumerService.java
@@ -2,7 +2,7 @@ package com.sparta.domain.consumer.service;
 
 import com.sparta.domain.consumer.dto.request.EditPasswordRequestDto;
 import com.sparta.domain.consumer.dto.request.EditProfileRequestDto;
-import com.sparta.domain.user.dto.UserResponseDto;
+import com.sparta.domain.user.dto.response.UserResponseDto;
 import com.sparta.domain.user.entity.User;
 import com.sparta.domain.user.repository.UserRepository;
 import com.sparta.global.exception.customException.UserException;
