+++ escape-core/src/main/java/com/sparta/domain/reaction/entity/Reaction.java
@@ -4,13 +4,14 @@ import com.sparta.domain.review.entity.Review;
 import com.sparta.domain.user.entity.User;
 import com.sparta.global.entity.TimeStamped;
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Reaction extends TimeStamped {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
