+++ escape-core/src/main/java/com/sparta/domain/review/entity/Review.java
@@ -6,6 +6,7 @@ import com.sparta.domain.theme.entity.Theme;
 import com.sparta.domain.user.entity.User;
 import com.sparta.global.entity.TimeStamped;
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,7 +16,7 @@ import java.util.List;
 
 @Entity
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Review extends TimeStamped {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
