+++ escape-core/src/main/java/com/sparta/domain/theme/entity/ThemeTime.java
@@ -2,6 +2,7 @@ package com.sparta.domain.theme.entity;
 
 
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,7 +12,7 @@ import java.time.LocalDateTime;
 
 @Entity
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ThemeTime {
 
     @Id
