--- escape-core/src/main/java/com/sparta/domain/user/dto/request/EmailVerificationRequestDto.java
@@ -1,16 +1,0 @@
-package com.sparta.domain.user.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class EmailVerificationRequestDto {
-    @NotBlank
-    private String email;
-
-    @NotNull
-    private String verificationCode;
-}
