+++ escape-core/src/main/java/com/sparta/domain/user/dto/response/UserResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.domain.user.dto;
+package com.sparta.domain.user.dto.response;
 
 import com.sparta.domain.user.entity.OAuthProvider;
 import com.sparta.domain.user.entity.User;
@@ -6,17 +6,13 @@ import com.sparta.domain.user.entity.UserStatus;
 import lombok.Getter;
 
 @Getter
-
 public class UserResponseDto {
-
     private Long id;
     private String email;
     private String name;
     private Long point;
     private UserStatus userStatus;
     private OAuthProvider oAuthProvider;
-
-
 
     public UserResponseDto(User user) {
         this.id = user.getId();
